move tidal tests to folder
